refactor stack cmds for ioctl
