+++ ioctl/cmd/action/action.go
@@ -76,8 +76,6 @@ func init() {
 	ActionCmd.AddCommand(actionReadCmd)
 	ActionCmd.AddCommand(actionClaimCmd)
 	ActionCmd.AddCommand(actionDepositCmd)
-	ActionCmd.AddCommand(actionStakeCmd)
-	ActionCmd.AddCommand(actionRestakeCmd)
 	ActionCmd.AddCommand(actionSendRawCmd)
 	ActionCmd.PersistentFlags().StringVar(&config.ReadConfig.Endpoint, "endpoint",
 		config.ReadConfig.Endpoint, "set endpoint for once")
