+++ ioctl/cmd/root.go
@@ -45,6 +45,7 @@ func init() {
 	RootCmd.AddCommand(update.UpdateCmd)
 	RootCmd.AddCommand(version.VersionCmd)
 	RootCmd.AddCommand(action.Xrc20Cmd)
+	RootCmd.AddCommand(action.StakeCmd)
 	RootCmd.PersistentFlags().StringVarP(&output.Format, "output-format", "o", "",
 		"output format")
 }
